ClientFolder/requirements.txt
python-dotenv==1.0.1
beautifulsoup4==4.10.0
psycopg2-binary==2.9.1
requests==2.26.0
Jinja2==3.1.3
flask==3.0.3